Add configuration management: create AppConfig class for environment variables and model settings
