+++ trainer/search_service.py
@@ -22,75 +22,13 @@ from transformers import AutoTokenizer, AutoModelForCausalLM
 from sentence_transformers import SentenceTransformer
 from peft import PeftModel, PeftConfig
 from flask import Flask, request, jsonify
+from .config import AppConfig
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 app = Flask(__name__)
 
-
-@dataclass
-class AppConfig:
-    REDIS_HOST: str = os.getenv("REDIS_HOST", "localhost")
-    REDIS_PORT: int = int(os.getenv("REDIS_PORT", 6379))
-    REDIS_PASSWORD: str = os.getenv("REDIS_PASSWORD", "")
-    MODEL_CACHE_SIZE: int = int(os.getenv("MODEL_CACHE_SIZE", 4))
-    MIN_SCORE: float = float(os.getenv("MIN_SCORE", 0.2))
-    AWS_ACCESS_KEY: str = os.getenv("AWS_ACCESS_KEY")
-    AWS_SECRET_KEY: str = os.getenv("AWS_SECRET_KEY")
-    S3_BUCKET: str = os.getenv("S3_BUCKET")
-    S3_REGION: str = os.getenv("AWS_REGION", "us-east-1")
-    AWS_SSL_VERIFY: bool = os.getenv("AWS_SSL_VERIFY", "true").lower() == "true"
-    AWS_ENDPOINT_URL: str = os.getenv("AWS_ENDPOINT_URL")
-    DEVICE: str = "cuda" if torch.cuda.is_available() else "cpu"
-    MAX_RETRIES: int = int(os.getenv("MAX_RETRIES", 3))
-    RETRY_DELAY: int = int(os.getenv("RETRY_DELAY", 1))
-
-    # Model configurations
-    AVAILABLE_LLM_MODELS = {
-        "distilgpt2-product": {
-            "name": "distilgpt2",
-            "description": "Fast, lightweight model for product search",
-            "peft_supported": True,
-        },
-        "all-minilm-l6": {
-            "name": "sentence-transformers/all-MiniLM-L6-v2",
-            "description": "Efficient embedding model for semantic search",
-            "peft_supported": False,
-        },
-    }
-    DEFAULT_MODEL = "distilgpt2-product"
-
-    # Simplified directory structure
-    BASE_MODEL_DIR: str = os.getenv("BASE_MODEL_DIR", "/app/models")
-    MODEL_CACHE_DIR: str = os.getenv("MODEL_CACHE_DIR", "/app/model_cache")
-    TRANSFORMER_CACHE: str = os.getenv(
-        "TRANSFORMER_CACHE", "/app/model_cache/transformers"
-    )
-    HF_HOME: str = os.getenv("HF_HOME", "/app/model_cache/huggingface")
-
-    @classmethod
-    def setup_cache_dirs(cls):
-        """Setup cache directories"""
-        directories = [
-            cls.BASE_MODEL_DIR,
-            cls.MODEL_CACHE_DIR,
-            cls.TRANSFORMER_CACHE,
-            cls.HF_HOME,
-            os.path.join(cls.HF_HOME, "datasets"),
-        ]
-
-        for directory in directories:
-            os.makedirs(directory, exist_ok=True)
-
-        os.environ.update(
-            {
-                "TORCH_HOME": cls.MODEL_CACHE_DIR,
-                "TRANSFORMERS_CACHE": cls.TRANSFORMER_CACHE,
-                "HF_HOME": cls.HF_HOME,
-                "HF_DATASETS_CACHE": os.path.join(cls.HF_HOME, "datasets"),
-            }
-        )
 
 class ModelCache:
     def __init__(self, max_size: int = 4):
